+++ src/maplibre.index.ts
@@ -1,3 +1,8 @@
-export * from './maplibre/canvas';
+import type { Map } from 'maplibre-gl';
+import { useMap as useMapGeneric } from './api/use-map';
+
 export * from './api/canvas-props';
 export * from './api/coordinates';
+export * from './maplibre/canvas';
+
+export const useMap = useMapGeneric<Map>;