+++ vite.config.ts
@@ -22,6 +22,7 @@ outDir = `${outDir}/${isES ? 'es' : 'cjs'}`;
 export default defineConfig({
   plugins: [react()],
   ...(!libMode
+    // story mode
     ? {
       base: '',
       resolve: {
@@ -31,8 +32,11 @@ export default defineConfig({
         }
       }
     }
+    // lib mode
     : {
+      publicDir: false,
       build: {
+        minify: false,
         lib: {
           entry,
           name: 'react-three-map',